ml-cluster/requirements.txt
# Flask y API
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# ML y Embeddings
sentence-transformers>=2.2.2
torch>=2.6.0
numpy>=1.26.0
scikit-learn>=1.3.0
hdbscan>=0.8.33

# Base de datos
supabase>=2.3.0
pgvector>=0.2.4
psycopg2-binary>=2.9.9

# Utilidades
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
